+++ cpp/lib/pairs.cpp
@@ -61,7 +61,7 @@ PairedSequenceMap pair_filter(MatchedSequenceMap matches, SequenceMap query)
     for (auto& dname_queries : matches) {
         PairedMatchesMap pairs;
         for (auto& qname_matches : dname_queries.second) {
-            for (auto& pair : flatten(qname_matches.second, query[qname_matches.first].size())) {
+            for (auto& pair : flatten(qname_matches.second, (u32)query[qname_matches.first].size())) {
                 if (ABS((s32)pair.dindex1 - (s32)pair.dindex2) <= (s32)query[qname_matches.first].size() - (s32)pair.length
                     || ABS((s32)pair.qindex1 - (s32)pair.qindex2) >= (s32)pair.length
                     || ABS((s32)pair.dindex1 - (s32)pair.dindex2) >= (s32)pair.length)
